moves all editable inputs to modal
