+++ src/app/characters/[id]/page.tsx
@@ -4,20 +4,14 @@ import { TraitsList } from '@/components/card/traits-list'
 import CharacterBuilderModal from '@/components/character-builder/character-builder'
 import { LabelsList } from '@/components/labels-list/labels-list'
 import { ParsedDescription } from '@/components/parsed-description/parsed-description'
-import { Ancestry, Attribute } from '@/models/db/ancestry'
-import { Background } from '@/models/db/background'
-import {
-  CharacterAncestry,
-  CharacterBackground,
-  CharacterEntity,
-} from '@/models/db/character-entity'
-import { Heritage } from '@/models/db/heritage'
+import { Attribute } from '@/models/db/ancestry'
+import { CharacterEntity } from '@/models/db/character-entity'
 import { PlayerCharacter } from '@/models/player-character'
 import { useDebounce } from '@/utils/debounce'
 import { roboto_serif } from '@/utils/fonts'
 import { getPlayerCharacter } from '@/utils/services/player-character-service'
 import { usePathname } from 'next/navigation'
-import React, { useEffect, useState } from 'react'
+import { useEffect, useState } from 'react'
 
 export default function CharacterPage() {
   const path: string[] = usePathname().split('/')
@@ -49,43 +43,14 @@ export default function CharacterPage() {
     saveEntity()
   })
 
-  const handleAncestryChange = (ancestryId: string) => {
-    character?.updateAncestry(ancestryId).then((val) => {
-      setCharacter(val)
-      debouncedRequest()
-    })
-  }
-
-  const handleAncestryEdit = (ancestry: CharacterAncestry) => {
-    let newCharacter: CharacterEntity = {
-      ...character!.getCharacter(),
-      ancestry: ancestry,
-    }
-    PlayerCharacter.build(newCharacter).then((val) => {
-      setCharacter(val)
-      debouncedRequest()
-    })
-  }
-
-  const handleBackgroundChange = (backgroundId: string) => {
-    character?.updateBackground(backgroundId).then((val) => {
-      setCharacter(val)
-      debouncedRequest()
-    })
-  }
-
-  const handleBackgroundEdit = (background: CharacterBackground) => {
-    let newCharacter: CharacterEntity = {
-      ...character!.getCharacter(),
-      background: background,
-    }
-    PlayerCharacter.build(newCharacter).then((val) => {
-      setCharacter(val)
-      debouncedRequest()
-    })
-  }
-
   const handleCharacterEdit = (char: CharacterEntity) => {
+    PlayerCharacter.build(char).then((val) => {
+      setCharacter(val)
+      debouncedRequest()
+    })
+  }
+
+  const handleClose = (char: CharacterEntity) => {
     PlayerCharacter.build(char).then((val) => {
       setCharacter(val)
       debouncedRequest()
@@ -96,42 +61,18 @@ export default function CharacterPage() {
     <div className={`h-full ${roboto_serif.className}`}>
       {character && (
         <>
-          <CharacterDisplay
-            character={character}
-            onEdit={handleCharacterEdit}
-          ></CharacterDisplay>
+          <CharacterDisplay character={character}></CharacterDisplay>
           <CharacterBuilderModal
             playerCharacter={character}
+            onClose={handleClose}
           ></CharacterBuilderModal>
-          <AncestryEdit
-            character={character}
-            onAncestryEdit={handleAncestryChange}
-            onEdit={handleAncestryEdit}
-          ></AncestryEdit>
-          <BackgroundEdit
-            character={character}
-            onBackgroundEdit={handleBackgroundChange}
-            onEdit={handleBackgroundEdit}
-          ></BackgroundEdit>
         </>
       )}
     </div>
   )
 }
 
-function CharacterDisplay({
-  character,
-  onEdit,
-}: {
-  character: PlayerCharacter
-  onEdit: (character: CharacterEntity) => void
-}) {
-  const updateName = (value: string) => {
-    let val: CharacterEntity = { ...character.getCharacter() }
-    val.name = value
-    onEdit(val)
-  }
-
+function CharacterDisplay({ character }: { character: PlayerCharacter }) {
   const languages = character.getLanguages()
   const senses = character.getSenses()
   const additionalFeatures = character.getAdditionalFeatures()
@@ -146,13 +87,7 @@ function CharacterDisplay({
           <div className="inline-flex gap-5 border border-stone-300 p-2">
             <div>
               <span>Name: </span>
-              <input
-                className="bg-stone-800"
-                value={character.getCharacter().name}
-                onChange={(e) => {
-                  updateName(e.target.value)
-                }}
-              ></input>
+              <span>{character.getCharacter().name}</span>
             </div>
             <div>
               <span>Level: </span>
@@ -312,262 +247,5 @@ function CharacterDisplay({
         <br />
       </div>
     )
-  )
-}
-
-// TODO separate out things like changing Ancestry with the choices from the Ancestry...
-function AncestryEdit({
-  character,
-  onAncestryEdit,
-  onEdit,
-}: {
-  character: PlayerCharacter
-  onAncestryEdit: (ancestryId: string) => void
-  onEdit: (val: CharacterAncestry) => void
-}) {
-  const [ancestries, setAncestries] = useState<Ancestry[]>([])
-  const [heritages, setHeritages] = useState<Heritage[]>([])
-
-  useEffect(() => {
-    fetch('http://localhost:3000/api/ancestries', {
-      cache: 'no-store',
-    })
-      .then((result) => result.json())
-      .then((ancestries) => {
-        setAncestries(ancestries)
-      })
-  }, [])
-
-  useEffect(() => {
-    fetch(
-      `http://localhost:3000/api/heritages?ancestry=${character.getAncestryName()}`,
-      {
-        cache: 'no-store',
-      }
-    )
-      .then((result) => result.json())
-      .then((heritages) => {
-        setHeritages(heritages)
-      })
-  }, [character.getCharacter().ancestry.id])
-
-  const updateAncestry = (value: string) => {
-    onAncestryEdit(value)
-  }
-
-  const updateAncestryAttributeMethod = () => {
-    let val: CharacterAncestry = { ...character.getCharacter().ancestry }
-    val.free_attribute = !val.free_attribute
-    onEdit(val)
-  }
-
-  const updateAncestryAttribute = (value: Attribute, index: number) => {
-    let val: CharacterAncestry = { ...character.getCharacter().ancestry }
-    val.attribute_boost_selections[index] = value
-    onEdit(val)
-  }
-
-  const updateAncestryLanguage = (value: string, index: number) => {
-    let val: CharacterAncestry = { ...character.getCharacter().ancestry }
-    val.language_selections[index] = value
-    onEdit(val)
-  }
-
-  const updateHeritage = (value: string) => {
-    let val: CharacterAncestry = { ...character.getCharacter().ancestry }
-    val.heritage_id = value
-    onEdit(val)
-  }
-
-  const languageChoices: (string | undefined)[] =
-    character?.getCharacter().ancestry.language_selections
-
-  return (
-    <div className="inline-flex gap-5 border border-stone-300 p-2 items-center">
-      <span>
-        <h2>Ancestry</h2>
-        <select
-          className="bg-stone-800"
-          value={character.getAncestryId()}
-          onChange={(e) => updateAncestry(e.target.value)}
-        >
-          {ancestries.map((ancestry) => (
-            <option
-              key={ancestry._id.toString()}
-              value={ancestry._id.toString()}
-            >
-              {ancestry.name}
-            </option>
-          ))}
-        </select>
-      </span>
-      <span>
-        {character && (
-          <label>
-            <input
-              className="bg-stone-800"
-              type="checkbox"
-              checked={character.getCharacter().ancestry.free_attribute}
-              onChange={(e) => updateAncestryAttributeMethod()}
-            />
-            Freely assign attributes?
-          </label>
-        )}
-
-        <h2>Attributes</h2>
-        {character &&
-          character
-            .getCharacter()
-            .ancestry.attribute_boost_selections.map(
-              (choice: any, i: number) => {
-                return (
-                  <React.Fragment key={i}>
-                    <select
-                      className="bg-stone-800 mr-2"
-                      value={choice ?? ''}
-                      onChange={(e) =>
-                        updateAncestryAttribute(e.target.value as Attribute, i)
-                      }
-                    >
-                      <option value={choice}>{choice}</option>
-                      {character
-                        .getAttributeChoices()
-                        .ancestry[i].map((attribute) => (
-                          <option key={attribute} value={attribute}>
-                            {attribute}
-                          </option>
-                        ))}
-                    </select>
-                  </React.Fragment>
-                )
-              }
-            )}
-      </span>
-      {languageChoices.length > 0 && (
-        <span>
-          <h2>Languages</h2>
-          {character &&
-            languageChoices.map((choice: any, i: number) => (
-              <React.Fragment key={i}>
-                <select
-                  className="bg-stone-800 mr-2"
-                  value={choice ?? ''}
-                  onChange={(e) => updateAncestryLanguage(e.target.value, i)}
-                >
-                  <option value={choice}>{choice}</option>
-                  {character.getLanguageChoices().ancestry.map((language) => (
-                    <option key={language} value={language}>
-                      {language}
-                    </option>
-                  ))}
-                </select>
-              </React.Fragment>
-            ))}
-        </span>
-      )}
-      <span>
-        <h2>Heritage</h2>
-        <select
-          className="bg-stone-800"
-          value={character.getCharacter().ancestry.heritage_id}
-          onChange={(e) => updateHeritage(e.target.value)}
-        >
-          <option value=""></option>
-          {heritages.map((heritage) => (
-            <option
-              key={heritage._id.toString()}
-              value={heritage._id.toString()}
-            >
-              {heritage.name}
-            </option>
-          ))}
-        </select>
-      </span>
-    </div>
-  )
-}
-
-// TODO separate out things like changing Ancestry with the choices from the Ancestry...
-function BackgroundEdit({
-  character,
-  onBackgroundEdit,
-  onEdit,
-}: {
-  character: PlayerCharacter
-  onBackgroundEdit: (backgroundId: string) => void
-  onEdit: (val: CharacterBackground) => void
-}) {
-  const [backgrounds, setBackgrounds] = useState<Background[]>([])
-
-  useEffect(() => {
-    fetch('http://localhost:3000/api/backgrounds', {
-      cache: 'no-store',
-    })
-      .then((result) => result.json())
-      .then((backgrounds) => {
-        setBackgrounds(backgrounds)
-      })
-  }, [])
-
-  const updateBackground = (value: string) => {
-    onBackgroundEdit(value)
-  }
-
-  const updateAttribute = (value: Attribute, index: number) => {
-    let val: CharacterBackground = { ...character.getCharacter().background }
-    val.attribute_boost_selections[index] = value
-    onEdit(val)
-  }
-
-  return (
-    <div className="inline-flex gap-5 border border-stone-300 p-2 items-center">
-      <span>
-        <h2>Background</h2>
-        <select
-          className="bg-stone-800"
-          value={character.getBackgroundId()}
-          onChange={(e) => updateBackground(e.target.value)}
-        >
-          {backgrounds.map((background) => (
-            <option
-              key={background._id.toString()}
-              value={background._id.toString()}
-            >
-              {background.name}
-            </option>
-          ))}
-        </select>
-      </span>
-      <span>
-        <h2>Attributes</h2>
-        {character?.getCharacter()?.background &&
-          character
-            .getCharacter()
-            .background.attribute_boost_selections.map(
-              (choice: any, i: number) => {
-                return (
-                  <React.Fragment key={i}>
-                    <select
-                      className="bg-stone-800 mr-2"
-                      value={choice ?? ''}
-                      onChange={(e) =>
-                        updateAttribute(e.target.value as Attribute, i)
-                      }
-                    >
-                      <option value={choice}>{choice}</option>
-                      {character
-                        .getAttributeChoices()
-                        .background[i].map((attribute) => (
-                          <option key={attribute} value={attribute}>
-                            {attribute}
-                          </option>
-                        ))}
-                    </select>
-                  </React.Fragment>
-                )
-              }
-            )}
-      </span>
-    </div>
   )
 }
